break inner classes into concrete classes
